Updated build.gradle to use newer versions of ForgeGradle, fix ghost creeper spawn, got rid of Illusion Creeper and made the fake ones a subtype using NBT, fixed EU creeper, re-added Ender, Spider, and Stone Creeper.

+++ src/main/java/net/lomeli/ec/entity/EntityGhostCreeper.java
@@ -1,8 +1,8 @@
 package net.lomeli.ec.entity;
 
-import net.lomeli.ec.lib.ECVars;
-
 import net.minecraft.world.World;
+
+import net.lomeli.ec.lib.ECVars;
 
 public class EntityGhostCreeper extends EntityBaseCreeper {
 
@@ -14,7 +14,7 @@ public class EntityGhostCreeper extends EntityBaseCreeper {
     @Override
     public void explosion(int power, boolean flag) {
         int exPower = this.explosionRadius * power;
-        this.worldObj.createExplosion(this, posX, posY, posZ, (float) exPower, false);
+        this.worldObj.createExplosion(this, posX, posY, posZ, exPower, false);
     }
 
 }
